test_email_notifications: Enable stream email notifs where appropriate.
When testing emails sent because of stream email notifications, the
right thing to do is to first enable the stream email notifications
setting for the test user (this setting is disabled by default).

Because `handle_missedmessage_emails` does not look at the user's
settings, these tests do not fail as of now. However, with the
notification settings recheck implementation coming up in the next
commit, that will not longer be true.